rib6: fix some highest bits handling
Without an explicit type, ip_xor gets promoted to a int.

../lib/rib/rte_rib6.c:348:26: runtime error:
	left shift of 130 by 24 places cannot be represented in type 'int'

Acked-by: Bruce Richardson <bruce.richardson@intel.com>
Acked-by: Vladimir Medvedkin <vladimir.medvedkin@intel.com>
Acked-by: Chengwen Feng <fengchengwen@huawei.com>

+++ lib/rib/rte_rib6.c
@@ -352,7 +352,7 @@ rte_rib6_insert(struct rte_rib6 *rib,
 		if (ip_xor == 0)
 			d += 8;
 		else {
-			d += rte_clz32(ip_xor << 24);
+			d += rte_clz32((uint32_t)ip_xor << 24);
 			break;
 		}
 	}
